requirements.txt
pandas
sqlalchemy
pymysql
streamlit
plotly
python-dotenv
